CI test orchestration workaround
